refactor(notification): even hander init
