+++ app/common/notification.go
@@ -10,6 +10,7 @@ import (
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/notification"
 	notificationadapter "github.com/openmeterio/openmeter/openmeter/notification/adapter"
+	"github.com/openmeterio/openmeter/openmeter/notification/eventhandler"
 	notificationservice "github.com/openmeterio/openmeter/openmeter/notification/service"
 	notificationwebhook "github.com/openmeterio/openmeter/openmeter/notification/webhook"
 	webhooknoop "github.com/openmeterio/openmeter/openmeter/notification/webhook/noop"
@@ -18,16 +19,16 @@ import (
 )
 
 var Notification = wire.NewSet(
+	NewNotificationAdapter,
 	NewNotificationService,
 	NewNotificationWebhookHandler,
+	NewNotificationEventHandler,
 )
 
-func NewNotificationService(
+func NewNotificationAdapter(
 	logger *slog.Logger,
 	db *entdb.Client,
-	webhook notificationwebhook.Handler,
-	featureConnector feature.FeatureConnector,
-) (notification.Service, error) {
+) (notification.Repository, error) {
 	adapter, err := notificationadapter.New(notificationadapter.Config{
 		Client: db,
 		Logger: logger,
@@ -36,6 +37,44 @@ func NewNotificationService(
 		return nil, fmt.Errorf("failed to initialize notification adapter: %w", err)
 	}
 
+	return adapter, nil
+}
+
+func NewNotificationEventHandler(
+	logger *slog.Logger,
+	adapter notification.Repository,
+	webhook notificationwebhook.Handler,
+) (notification.EventHandler, func(), error) {
+	closeFn := func() {}
+
+	eventHandler, err := eventhandler.New(eventhandler.Config{
+		Repository: adapter,
+		Webhook:    webhook,
+		Logger:     logger,
+	})
+	if err != nil {
+		return nil, closeFn, fmt.Errorf("failed to initialize notification event handler: %w", err)
+	}
+
+	if err = eventHandler.Start(); err != nil {
+		return nil, closeFn, fmt.Errorf("failed to initialize notification event handler: %w", err)
+	}
+
+	closeFn = func() {
+		if err = eventHandler.Close(); err != nil {
+			logger.Error("failed to close notification event handler", "error", err)
+		}
+	}
+
+	return eventHandler, closeFn, nil
+}
+
+func NewNotificationService(
+	logger *slog.Logger,
+	adapter notification.Repository,
+	webhook notificationwebhook.Handler,
+	featureConnector feature.FeatureConnector,
+) (notification.Service, error) {
 	notificationService, err := notificationservice.New(notificationservice.Config{
 		Adapter:          adapter,
 		Webhook:          webhook,
